Moved the code from Program.cs to Application.cs enable tests in this part of application.
