+++ src/Contact.Api/Application.cs
@@ -1,0 +1,53 @@
+﻿using Contacts.Application.Contexts;
+using Contacts.Infrastructure.DI;
+using Microsoft.EntityFrameworkCore;
+
+namespace Contacts.Api
+{
+    public class Application
+    {
+        public void Run(string[] args)
+        {
+            var builder = WebApplication.CreateBuilder(args);
+
+            // Add services to the container.
+
+            builder.Configuration.AddJsonFile("appsettings.json", optional: false, reloadOnChange: true);
+
+            builder.Services.AddSingleton<DbInitializer>();
+
+            builder.Services.AddDbContext<AppDbContext>(options =>
+                options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
+
+            builder.Services.AddAutoRegister<AppDbContext>();
+
+            builder.Services.AddControllers();
+            // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+            builder.Services.AddEndpointsApiExplorer();
+            builder.Services.AddSwaggerGen();
+
+            var app = builder.Build();
+
+            using (var scope = app.Services.CreateScope())
+            {
+                var migrationRunner = scope.ServiceProvider.GetRequiredService<DbInitializer>();
+                migrationRunner.Initialize();
+            }
+
+            // Configure the HTTP request pipeline.
+            if (app.Environment.IsDevelopment())
+            {
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+
+            app.UseHttpsRedirection();
+
+            app.UseAuthorization();
+
+            app.MapControllers();
+
+            app.RunAsync();
+        }
+    }
+}
